requirements.txt
Django==3.2.2
requests==2.25.1
djangorestframework==3.12.4
django-rest-auth==0.9.5
# Filtros
django-filter==22.1

matplotlib==3.4.3
# Para trabajar con imagenes
Pillow==8.4.0

# Swagger:
uritemplate==3.0.1
pyyaml==5.4.1

# Drivers para trabajar con Postgres en linux:

psycopg2-binary>=2.8



